lncli: move openChannel command to new file
This is a pure code move!
